requirements.txt
# -f https://download.pytorch.org/whl/lts/1.10/torch_lts.html
# gdown
# Cython
# matplotlib
# numpy
# opencv-contrib-python
# pillow
# PyYAML
# scipy
# tqdm
# flask
# flask-cors
# gunicorn
# torch
# torchvision
# pandas
# seaborn



# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0+cpu,!=1.12.0+cpu
torchvision>=0.8.1+cpu,!=0.13.0+cpu
tqdm>=4.41.0
protobuf<4.21.3
flask
flask-cors

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0